+++ gwcelery/tests/test_tasks_gracedb.py
@@ -104,6 +104,12 @@ def test_get_superevent(mock_gracedb):
 
 
 @patch('gwcelery.tasks.gracedb.client', autospec=rest.GraceDb)
+def test_get_superevents(mock_gracedb):
+    gracedb.get_superevents('query')
+    mock_gracedb.superevents.assert_called_once_with('query', orderby='t_0')
+
+
+@patch('gwcelery.tasks.gracedb.client', autospec=rest.GraceDb)
 def test_upload(mock_gracedb):
     # Run function under test.
     gracedb.upload('filecontents', 'filename', 'graceid', 'message', 'tags')